requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
pyserial==3.5
requests==2.32.3